experiments/doc_assistant/faiss_wip/requirements.txt
streamlit
openai
python-dotenv
PyMuPDF
pymupdf4llm
jieba

# MinerU 依赖
modelscope
magic-pdf[full]
--extra-index-url https://wheels.myhloli.com
--index-url https://mirrors.aliyun.com/pypi/simple